requirements.txt
pygame
SpeechRecognition
pywhatkit
google-genai
edge-tts
pywin32
python-dotenv
